passage à spring data rest

+++ BibApp/src/main/java/dumaya/dev/BibApp/repository/ReferenceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReferenceRepository extends JpaRepository<Reference, Long> {
+public interface ReferenceRepository extends JpaRepository<Reference, Integer> {
 
     Reference findById(int id);
